only authorized users can delete a user

+++ app/Http/Controllers/Admin/UserController.php
@@ -56,6 +56,7 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        $this->authorize('delete-user');
         $user->delete();
         return response()->json([
             'message' => ('User successfully deleted')
